Add Clone and 3 new fn's to bdd.rs
1) Add derive clone for bdd.rs

A BDD (actually a CRHS eq) can now be cloned. As for a System, this
should be done with considerable care, as cloning a CRHS equation which
already uses more than half of the available memory probably will lead
to the program crashing. This is therefore currently hidden behind the
"differential" feature.

2) Add 3 fn's to bdd.rs, make some fn's pub

Three functions are added: level, set_lhs_level_from_vob and
get_lhs_level_from_vob. The two latter fn's deal with the LHS of a
specific level of the CRHS eq, while the former yields a pointer to a
RHS Level, if it exists.

In addition, some functions were made public. These additions and
changes are made as part of an ongoing experimentation of
exposing more of the underlying API to outside libs/bins, in an attempt
to avoid CRUSH itself to explode with functions for various needs.
(Functions that would by logic rather belong to a different lib/bin. We
will see an example of this later, when some of the ops needed for PF
are migrated in. I.e. some of those fn's probably should be in a
different lib, but due to the need of functionality only present in
CRUSH, and not easily exposed to other libs, they are implemented
in CRUSH. I'd like to avoid this as much as possible).