Fix a bug when deducing .bdd in-file name
There was missing an underscore in the deduction algorithm when
loading a .bdd from file. This is now fixed.

This also implies that it is not the same algorithm making the file
names when saving to file as when loading from file, which strikes me
as odd. TODO: This should be unified.